refactor: Introduce generator class to handle global flow
